feat(Products): Setup products model
