+++ superpool/api/core/catalog/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class CatalogConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "core.catalog"
